Revert "migration to typecscript (2)"
This reverts commit 2d03cf2fd17732208a10861c46ae56fbd0b3334a.
