+++ src/components/ui/Accordion/shards/AccordionItem.tsx
@@ -5,10 +5,10 @@ import {customClassSwitcher} from '~/core';
 interface AccordionItemProps {
     children: React.ReactNode;
     customItemClass?: string;
-    value?: number;
+    value: string;
 }
 
-const AccordionItem: React.FC<AccordionItemProps> = ({children, value, customItemClass=''}) => {
+const AccordionItem: React.FC<AccordionItemProps> = ({children, value='', customItemClass=''}) => {
     const rootClass = customClassSwitcher(customItemClass, 'Accordion');
     return (
         <div className={`${rootClass}-item`}>
