+++ src/components/ui/Code/Code.js
@@ -1,11 +1,7 @@
 'use client';
 import React from 'react';
 
-interface CodeProps {
-    children: React.ReactNode;
-}
-
-const Code = ({children}: CodeProps) => {
+const Code = ({children}) => {
     return <code className='rui-code-root'>
         {children}
     </code>;
