+++ src/components/ui/Em/Em.js
@@ -4,14 +4,7 @@ import {customClassSwitcher} from '~/core';
 
 const COMPONENT_NAME = 'Em';
 
-interface EmProps {
-    children: React.ReactNode;
-    customRootClass?: string;
-    className?: string;
-    props: Record<string, any>[]
-}
-
-const Em = ({children, customRootClass, className, ...props}: EmProps) => {
+const Em = ({children, customRootClass = '', className = '', ...props}) => {
     const rootClass = customClassSwitcher(customRootClass, COMPONENT_NAME);
     return <em className={`${rootClass} ${className}`} {...props}>
         {children}
