+++ src/components/ui/Quote/Quote.js
@@ -3,15 +3,7 @@ import React from 'react';
 import {customClassSwitcher} from '~/core';
 
 const COMPONENT_NAME = 'Quote';
-
-interface QuoteProps {
-    children: React.ReactNode;
-    customRootClass?: string;
-    className?: string;
-    props?: Record<string, any>[]
-}
-
-const Quote = ({children, customRootClass, className, ...props}: QuoteProps) => {
+const Quote = ({children, customRootClass = '', className = '', ...props}) => {
     const rootClass = customClassSwitcher(customRootClass, COMPONENT_NAME);
     return <q className={`${rootClass} ${className}`} {...props}>{children}</q>;
 };
